Item shown in ActivityProductDetails

+++ app/src/main/java/com/team12/ClassAddProduct.java
@@ -2,19 +2,19 @@ package com.team12;
 
 public class ClassAddProduct {
     String name, description, image, sellerName;
-    int price, sellerId;
+    long price, sellerId;
 
     public ClassAddProduct(){
 
     }
-    public ClassAddProduct(String name, String description, String image, int price, int sellerId) {
+    public ClassAddProduct(String name, String description, String image, long price, long sellerId) {
         this.name = name;
         this.description = description;
         this.image = image;
         this.price = price;
         this.sellerId = sellerId;
     }
-    public ClassAddProduct(String name, String description, String image, int price, int sellerId, String sellerName) {
+    public ClassAddProduct(String name, String description, String image, long price, long sellerId, String sellerName) {
         this.name = name;
         this.description = description;
         this.image = image;
@@ -47,7 +47,7 @@ public class ClassAddProduct {
         this.image = image;
     }
 
-    public int getPrice() {
+    public long getPrice() {
         return price;
     }
 
@@ -55,7 +55,7 @@ public class ClassAddProduct {
         this.price = price;
     }
 
-    public int getSellerId() {
+    public long getSellerId() {
         return sellerId;
     }
